backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
click==8.3.0
exceptiongroup==1.3.0
fastapi==0.121.2
h11==0.16.0
httptools==0.7.1
httpx==0.28.1  # For downloading files from Stream CDN
idna==3.11
pydantic==2.12.4
pydantic_core==2.41.5
python-dotenv==1.2.1
PyYAML==6.0.3
groq==0.34.1
sniffio==1.3.1
starlette==0.49.3
typing-inspection==0.4.2
typing_extensions==4.15.0
stream-chat==4.27.0
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1

# Knowledge base / RAG dependencies (FREE!)
pinecone-client==5.0.1
sentence-transformers==3.3.1

# File processing dependencies
python-multipart==0.0.18  # For file uploads
PyPDF2==3.0.1  # For PDF text extraction
python-docx==1.1.2  # For Word document extraction
openpyxl==3.1.5  # For Excel files
python-pptx==1.0.2  # For PowerPoint files
